requirements.txt
django>=6.0,<7.0
djangorestframework>=3.15,<4.0
gunicorn>=23.0.0
psycopg[binary]>=3.2
